Rewrited sgfanalyze
