+++ sgfanalyze.py
@@ -1,40 +1,99 @@
-import datetime
 import hashlib
 import os
 import pickle
-import re
 import sys
 import time
 import traceback
-
 import config
+import re
+import datetime
+from sgftools import regex
 import sgftools.utils as utils
 from sgftools import gotools, annotations, progressbar, sgflib
-from sgftools import regex
 from sgftools.leela import Leela
+
+
+def calculate_tasks_left(sgf_file, request_analyze, request_variations):
+    """
+    Calculate tasks left
+    """
+    sgf_cursor = sgf_file.cursor()
+    move_num = 0
+    analyze_tasks = 0
+    variations_tasks = 0
+
+    # Look through every node
+    while not sgf_cursor.atEnd:
+        sgf_cursor.next()
+        analysis_mode = None
+
+        # Conditions
+        is_default_analysis = args.analyze_start <= move_num <= args.analyze_end
+        is_analyze_request = move_num in request_analyze
+        is_prev_analyze_request = (move_num - 1) in request_analyze
+        is_vars_request = move_num in request_variations
+        is_prev_vars_request = (move_num - 1) in request_variations
+
+        if is_vars_request:
+            analyze_tasks += 1
+            variations_tasks += 1
+        elif any([is_default_analysis, is_analyze_request, is_prev_analyze_request, is_prev_vars_request]):
+            analyze_tasks += 1
+
+        move_num += 1
+
+    return analyze_tasks, variations_tasks
 
 
 def add_moves_to_leela(cursor, leela):
     this_move = None
 
-    if 'W' in cursor.node.keys():
-        this_move = cursor.node['W'].data[0]
-        leela.add_move('white', this_move)
-
-    if 'B' in cursor.node.keys():
-        this_move = cursor.node['B'].data[0]
+    # Get moves nodes
+    node_black_move = cursor.node.get('B')
+    node_white_move = cursor.node.get('W')
+    node_black_stones = cursor.node.get('AB')
+    node_white_stones = cursor.node.get('AW')
+    # Store commands to add black or white moves
+    if node_black_move:
+        this_move = node_black_move.data[0]
         leela.add_move('black', this_move)
 
-    # SGF commands to add black or white stones, often used for setting up handicap and such
-    if 'AB' in cursor.node.keys():
-        for move in cursor.node['AB'].data:
+    elif node_white_move:
+        this_move = node_white_move.data[0]
+        leela.add_move('white', this_move)
+
+    # Store commands to add black or white stones, often used for setting up handicap and such
+    if node_black_stones:
+        for move in node_black_stones.data:
             leela.add_move('black', move)
 
-    if 'AW' in cursor.node.keys():
-        for move in cursor.node['AW'].data:
+    if node_white_stones:
+        for move in node_white_stones.data:
             leela.add_move('white', move)
 
     return this_move
+
+
+def next_move_pos(cursor):
+    """
+    Go to next node and look what move was played
+    :return: SGF-like move string
+    """
+    next_move = None
+
+    if not cursor.atEnd:
+        cursor.next()
+        node_white_move = cursor.node.get('W')
+        node_black_move = cursor.node.get('B')
+
+        if node_white_move:
+            next_move = node_white_move.data[0]
+        elif node_black_move:
+            next_move = node_black_move.data[0]
+
+        cursor.previous()
+
+    return next_move
 
 
 def retry_analysis(fn):
@@ -53,15 +112,17 @@ def retry_analysis(fn):
 
 
 @retry_analysis
-def do_analyze(leela, base_dir, verbosity, seconds_per_search):
+def do_analyze(leela, base_dir, seconds_per_search):
     ckpt_hash = 'analyze_' + leela.history_hash() + "_" + str(seconds_per_search) + "sec"
     ckpt_fn = os.path.join(base_dir, ckpt_hash)
-    # if verbosity > 2:
-    #     print("Looking for checkpoint file: %s" % ckpt_fn, file=sys.stderr)
 
-    if os.path.exists(ckpt_fn) and not config.defaults['skip_checkpoints']:
-        if verbosity > 1:
+    if args.verbosity > 2:
+        print("Looking for checkpoint file: %s" % ckpt_fn, file=sys.stderr)
+
+    if os.path.exists(ckpt_fn):
+        if args.verbosity > 2:
             print("Loading checkpoint file: %s" % ckpt_fn, file=sys.stderr)
+
         with open(ckpt_fn, 'rb') as ckpt_file:
             stats, move_list = pickle.load(ckpt_file)
             ckpt_file.close()
@@ -76,21 +137,34 @@ def do_analyze(leela, base_dir, verbosity, seconds_per_search):
     return stats, move_list
 
 
-# move_list is from a call to do_analyze
-# Iteratively expands a tree of moves by expanding on the leaf with the highest "probability of reaching".
-def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_dir, args):
-    nodes_per_variation = args.nodes_per_variation
-    verbosity = args.verbosity
+def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_dir):
+    """
+    Analyze variations and add them to SGF
+    """
 
+    # Do not analyze variations if Leela still in opening book
     if 'bookmoves' in stats or len(move_list) <= 0:
         return None
 
     rootcolor = leela.whose_turn()
+
+    # Initiate tree and leaves of SGF
     leaves = []
-    tree = {"children": [], "is_root": True, "history": [], "explored": False, "prob": 1.0, "stats": stats,
-            "move_list": move_list, "color": rootcolor}
+    tree = {
+        "children": [],
+        "is_root": True,
+        "history": [],
+        "explored": False,
+        "prob": 1.0,
+        "stats": stats,
+        "move_list": move_list,
+        "color": rootcolor
+    }
 
     def expand(node, stats, move_list):
+        """
+        Expand node with variations
+        """
         assert node["color"] in ['white', 'black']
 
         def child_prob_raw(i, move):
@@ -106,6 +180,7 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
             return child_prob_raw(i, move) / probsum
 
         probsum = 0.0
+
         for (i, move) in enumerate(move_list):
             probsum += child_prob_raw(i, move)
 
@@ -119,8 +194,16 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
             subhistory.append(move["pos"])
             prob = node["prob"] * child_prob(i, move)
             clr = "white" if node["color"] == "black" else "black"
-            child = {"children": [], "is_root": False, "history": subhistory, "explored": False, "prob": prob,
-                     "stats": {}, "move_list": [], "color": clr}
+            child = {
+                "children": [],
+                "is_root": False,
+                "history": subhistory,
+                "explored": False,
+                "prob": prob,
+                "stats": {},
+                "move_list": [],
+                "color": clr
+            }
             node["children"].append(child)
             leaves.append(child)
 
@@ -134,16 +217,21 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
                 break
 
     def search(node):
+        """
+        Analyze variation
+        """
         for mv in node["history"]:
             leela.add_move(leela.whose_turn(), mv)
-        stats, move_list = do_analyze(leela, base_dir, verbosity, args.variations_time)
+
+        stats, move_list = do_analyze(leela, base_dir, args.variations_time)
         expand(node, stats, move_list)
 
         for mv in node["history"]:
             leela.pop_move()
 
     expand(tree, stats, move_list)
-    for i in range(nodes_per_variation):
+
+    for i in range(args.nodes_per_variation):
         if len(leaves) > 0:
             node = max(leaves, key=(lambda n: n["prob"]))
             search(node)
@@ -165,10 +253,13 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
             cursor.next(len(cursor.children) - 1)
 
     def record(node):
+        """
+        Write node to SGF
+        """
         if not node["is_root"]:
             annotations.annotate_sgf(cursor,
                                      annotations.format_winrate(node["stats"], node["move_list"], board_size, None),
-                                     [], [])
+                                     None, None)
             move_list_to_display = []
 
             # Only display info for the principal variation or for lines that have been explored.
@@ -190,10 +281,12 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
                     advance(cursor, node["color"], child["history"][-1])
                     record(child)
                     cursor.previous()
+
                 # Only show variations for the principal line, to prevent info overload
                 elif i == 0:
                     pv = node["move_list"][i]["pv"]
                     color = node["color"]
+
                     num_to_show = min(len(pv), max(1, len(pv) * 2 / 3 - 1))
 
                     if args.num_to_show is not None:
@@ -209,38 +302,6 @@ def do_variations(cursor, leela, stats, move_list, board_size, game_move, base_d
     record(tree)
 
 
-def calculate_tasks_left(sgf, comment_requests_analyze, comment_requests_variations):
-    cursor = sgf.cursor()
-    move_num = 0
-    analyze_tasks = 0
-    variations_tasks = 0
-    while not cursor.atEnd:
-        cursor.next()
-
-        analysis_mode = None
-        if args.analyze_start <= move_num <= args.analyze_end:
-            analysis_mode = 'analyze'
-
-        if move_num in comment_requests_analyze or (move_num - 1) in comment_requests_analyze or (
-                    move_num - 1) in comment_requests_variations:
-            analysis_mode = 'analyze'
-
-        if move_num in comment_requests_variations:
-            analysis_mode = 'variations'
-
-        if analysis_mode == 'analyze':
-            analyze_tasks += 1
-        elif analysis_mode == 'variations':
-            analyze_tasks += 1
-            variations_tasks += 1
-
-        move_num += 1
-    return analyze_tasks, variations_tasks
-
-
-default_analyze_thresh = 0.010
-default_var_thresh = 0.010
-
 if __name__ == '__main__':
 
     time_start = datetime.datetime.now()
@@ -254,31 +315,30 @@ if __name__ == '__main__':
 
     sgf_fn = args.SGF_FILE
 
-    # if no file name to save analyze results provided - it will use original source file with concat 'analyzed'
     if not args.save_to_file:
         # FIXME: possible bug with folders which contain "." in their names
         args.save_to_file = args.SGF_FILE.split('.')[0] + '_analyzed.sgf'
 
     if not os.path.exists(sgf_fn):
-        config.parser.error("No such file: %s" % (sgf_fn))
+        config.parser.error("No such file: %s" % sgf_fn)
     sgf = gotools.import_sgf(sgf_fn)
 
     RESTART_COUNT = args.restarts
 
+    # Create checkpoints directory f not exists
     if not os.path.exists(args.ckpt_dir):
         os.mkdir(args.ckpt_dir)
 
+    # Create base hash and directory for analysis
     base_hash = hashlib.md5(os.path.abspath(sgf_fn).encode()).hexdigest()
     base_dir = os.path.join(args.ckpt_dir, base_hash)
 
+    # Create base if not exists
     if not os.path.exists(base_dir):
         os.mkdir(base_dir)
 
     if args.verbosity > 1:
         print("Checkpoint dir: %s" % base_dir, file=sys.stderr)
-
-    comment_requests_analyze = {}  # will contain move numbers that should be analyzed
-    comment_requests_variations = {}  # will contain move numbers that should be analyzed with variations
 
     # Set up SGF cursor
     cursor = sgf.cursor()
@@ -321,13 +381,16 @@ if __name__ == '__main__':
                 old_komi, komi, handicap_stone_count), file=sys.stderr)
     else:
         if is_handicap_game:
-            komi = 0.5
+            komi = 0.5 if is_japanese_rules else 0.5 + handicap_stone_count
         else:
             komi = 6.5 if is_japanese_rules else 7.5
         print("Warning: Komi not specified, assuming %.1f" % komi, file=sys.stderr)
 
-    # First loop for comments parsing
+    # COLLECT MOVES REQUESTED FOR ANALYSIS AND VARIATIONS
+    comment_requests_analyze = {}  # will contain moves requested for analysis
+    comment_requests_variations = {}  # will contain moves requested for varioations
     move_num = -1
+
     while not cursor.atEnd:
 
         # Go to next node and increment move_num
@@ -339,19 +402,21 @@ if __name__ == '__main__':
         # Store moves, requested for analysis and variations
         if node_comment:
             match = re.match(regex.comment_regex, node_comment.data[0])
-            if match:
-                if 'analyze' in match.group('node_comment'):
-                    comment_requests_analyze[move_num] = True
 
-                if 'variations' in match.group('node_comment'):
-                    comment_requests_analyze[move_num] = True
-                    comment_requests_variations[move_num] = True
+            if 'analyze' in match.group('node_comment'):
+                comment_requests_analyze[move_num] = True
+
+            if 'variations' in match.group('node_comment'):
+                comment_requests_analyze[move_num] = True
+                comment_requests_variations[move_num] = True
 
             # Wipe comments is needed
             if args.wipe_comments:
                 node_comment.data[0] = ""
 
-    # Calculate initial tasks
+    # END OF SGF-FILE REACHED
+
+    # Count initial analyze and variations tasks
     (analyze_tasks_initial, variations_tasks_initial) = calculate_tasks_left(sgf, comment_requests_analyze,
                                                                              comment_requests_variations)
     variations_task_probability = 1.0 / (1.0 + args.variations_threshold * 100.0)
@@ -361,6 +426,9 @@ if __name__ == '__main__':
 
 
     def approx_tasks_done():
+        """
+        Calculate approximate tasks done
+        """
         return (
             analyze_tasks_initial_done +
             (variations_tasks_done * args.nodes_per_variation)
@@ -368,6 +436,9 @@ if __name__ == '__main__':
 
 
     def approx_tasks_max():
+        """
+        Calculate approximate tasks max
+        """
         return (
             (analyze_tasks_initial - analyze_tasks_initial_done) *
             (1 + variations_task_probability * args.nodes_per_variation) +
@@ -378,22 +449,25 @@ if __name__ == '__main__':
 
     transform_winrate = utils.winrate_transformer(config.defaults['stdev'], args.verbosity)
 
-    analyze_threshold = transform_winrate(0.5 + 0.5 * args.analyze_threshold) - \
-                        transform_winrate(0.5 - 0.5 * args.analyze_threshold)
+    analyze_threshold = transform_winrate(0.5 + 0.5 * args.analyze_threshold) - transform_winrate(
+        0.5 - 0.5 * args.analyze_threshold)
+    variations_threshold = transform_winrate(0.5 + 0.5 * args.variations_threshold) - transform_winrate(
+        0.5 - 0.5 * args.variations_threshold)
 
-    variations_threshold = transform_winrate(0.5 + 0.5 * args.variations_threshold) - \
-                           transform_winrate(0.5 - 0.5 * args.variations_threshold)
-
-    print("Executing approx %.0f analysis steps" % approx_tasks_max(), file=sys.stderr)
+    print("Executing approx %d analysis steps" % approx_tasks_max(), file=sys.stderr)
 
     progress_bar = progressbar.ProgressBar(max_value=approx_tasks_max())
     progress_bar.start()
 
 
     def refresh_progress_bar():
+        """
+        Refresh progress bar
+        """
         progress_bar.update(approx_tasks_done(), approx_tasks_max())
 
 
+    # Create Leela
     leela = Leela(board_size=board_size,
                   executable=args.executable,
                   is_handicap_game=is_handicap_game,
@@ -407,34 +481,44 @@ if __name__ == '__main__':
     needs_variations = {}
 
     try:
-        move_num = -1
         cursor = sgf.cursor()
+        move_num = -1
         prev_stats = {}
         prev_move_list = []
-        has_prev = False
+        has_previous = False
 
+        # Start Leela process
         leela.start()
         add_moves_to_leela(cursor, leela)
 
-        # analyze main line, without variations
+        # ANALYZE MAIN LINE AND COLLECT WINRATES
         while not cursor.atEnd:
             cursor.next()
             move_num += 1
+            print([v for v in needs_variations])
+
+            # Add move to history and get whose turn is now
             this_move = add_moves_to_leela(cursor, leela)
             current_player = leela.whose_turn()
             prev_player = "white" if current_player == "black" else "black"
 
-            if ((args.analyze_start <= move_num <= args.analyze_end) or
-                    (move_num in comment_requests_analyze) or
-                    ((move_num - 1) in comment_requests_analyze) or
-                    (move_num in comment_requests_variations) or
-                    ((move_num - 1) in comment_requests_variations)):
+            # Conditions to start analysis
+            is_default_analysis = args.analyze_start <= move_num <= args.analyze_end
+            is_analyze_request = move_num in comment_requests_analyze
+            is_prev_analyze_request = (move_num - 1) in comment_requests_analyze
+            is_vars_request = move_num in comment_requests_variations
+            is_prev_vars_request = (move_num - 1) in comment_requests_variations
 
-                stats, move_list = do_analyze(leela, base_dir, args.verbosity, args.analyze_time)
+            # START MAIN LINE ANALYSIS
+            if any([is_default_analysis, is_analyze_request, is_prev_analyze_request,
+                    is_vars_request, is_prev_vars_request]):
+                stats, move_list = do_analyze(leela, base_dir, args.analyze_time)
 
+                # Store winrate of black player after played move
                 if 'winrate' in stats and stats['visits'] > 100:
                     collected_winrates[move_num] = (current_player, stats['winrate'])
 
+                # Store best move winrate
                 if len(move_list) > 0 and 'winrate' in move_list[0]:
                     collected_best_moves[move_num] = move_list[0]['pos']
                     collected_best_move_winrates[move_num] = move_list[0]['winrate']
@@ -442,114 +526,111 @@ if __name__ == '__main__':
                 delta = 0.0
                 transdelta = 0.0
 
+                # If winrate of current and previous moves exist, calculate delta
                 if 'winrate' in stats and (move_num - 1) in collected_best_moves:
+                    # Calculate delta and transdelta if played move is not Leela's "best move"
                     if this_move != collected_best_moves[move_num - 1]:
                         delta = stats['winrate'] - collected_best_move_winrates[move_num - 1]
-                        delta = min(0.0, (-delta if leela.whose_turn() == "black" else delta))
-                        transdelta = transform_winrate(stats['winrate']) - \
-                                     transform_winrate(collected_best_move_winrates[move_num - 1])
+                        delta = min(0.0, (-delta if leela.whose_turn() == "black" else delta))  # adjust delta <= 0
+                        transdelta = transform_winrate(stats['winrate']) - transform_winrate(
+                            collected_best_move_winrates[move_num - 1])
                         transdelta = min(0.0, (-transdelta if leela.whose_turn() == "black" else transdelta))
 
                     if transdelta <= -analyze_threshold:
                         (delta_comment, delta_lb_values) = annotations.format_delta_info(delta, transdelta, stats,
                                                                                          this_move, board_size)
-                        annotations.annotate_sgf(cursor, delta_comment, delta_lb_values, [])
+                        annotations.annotate_sgf(cursor, delta_comment, delta_lb_values)
 
-                if has_prev and (transdelta <= -variations_threshold or (move_num - 1) in comment_requests_variations):
-                    if not (args.skip_white and prev_player == "white") and not (
-                                args.skip_black and prev_player == "black"):
-                        needs_variations[move_num - 1] = (prev_stats, prev_move_list)
+                is_reached_analyze_thresh = transdelta <= -analyze_threshold
+                is_reached_vars_thresh = transdelta <= -variations_threshold
+                is_skipped_for_vars = (args.skip_white and prev_player == "white") or \
+                                      (args.skip_black and prev_player == "black")
 
-                        if (move_num - 1) not in comment_requests_variations:
-                            variations_tasks += 1
+                # If previous move was analyzed and current move is not skipped,
+                # store moves that should be analyzed with variations
+                if has_previous and not is_skipped_for_vars and (is_reached_vars_thresh or is_prev_vars_request):
+                    needs_variations[move_num - 1] = (prev_stats, prev_move_list)
 
-                next_game_move = None
+                    # Increment counter if previous move is not comment requested
+                    if not is_prev_vars_request:
+                        variations_tasks += 1
 
-                if not cursor.atEnd:
+                # Look where the next move was played
+                next_game_move = next_move_pos(cursor)
+
+                # Add winrate comment
+                winrate_comment = annotations.format_winrate(stats, move_list, board_size, next_game_move)
+                annotations.annotate_sgf(cursor, winrate_comment)
+
+                # Add comment with analysis to previous move
+                if has_previous and (is_prev_analyze_request or is_prev_vars_request or is_reached_analyze_thresh) and \
+                        not is_skipped_for_vars:
+                    (analysis_comment, lb_values, tr_values) = annotations.format_analysis(
+                        prev_stats, prev_move_list, this_move)
+                    cursor.previous()
+                    annotations.annotate_sgf(cursor, analysis_comment, lb_values, tr_values)
                     cursor.next()
 
-                    if 'W' in cursor.node.keys():
-                        next_game_move = cursor.node['W'].data[0]
-
-                    if 'B' in cursor.node.keys():
-                        next_game_move = cursor.node['B'].data[0]
-
-                    cursor.previous()
-
-                annotations.annotate_sgf(cursor,
-                                         annotations.format_winrate(stats, move_list, board_size, next_game_move),
-                                         [], [])
-
-                if has_prev and ((move_num - 1) in comment_requests_analyze or (
-                            move_num - 1) in comment_requests_variations or transdelta <= -analyze_threshold):
-                    if not (args.skip_white and prev_player == "white") and not (
-                                args.skip_black and prev_player == "black"):
-                        (analysis_comment, lb_values, tr_values) = annotations.format_analysis(prev_stats,
-                                                                                               prev_move_list,
-                                                                                               this_move)
-                        cursor.previous()
-                        # adding comment to sgf with suggested alternative variations
-                        annotations.annotate_sgf(cursor, analysis_comment, lb_values, tr_values)
-                        cursor.next()
-
+                # Save move information for next move analysis
                 prev_stats = stats
                 prev_move_list = move_list
-                has_prev = True
-                analyze_tasks_initial_done += 1
+                has_previous = True
+                analyze_tasks_initial_done += 1  # count completed task
 
-                # save to file results with analyzing main line
+                # Update sgf-file and refresh progressbar
                 utils.write_to_file(args.save_to_file, 'w', sgf)
-
                 refresh_progress_bar()
 
-                # until now analyze of main line, without sub-variations
-
+            # SKIP MAIN LINE ANALYSIS
             else:
                 prev_stats = {}
                 prev_move_list = []
-                has_prev = False
+                has_previous = False
 
+                # END MAIN LINE ANALYSIS
+
+        # Stop leela process and clear history
         leela.stop()
         leela.clear_history()
 
+        # Update sgf-file
+        utils.write_to_file(args.save_to_file, 'w', sgf)
+
+        # Save winrate graph to file
         if args.win_graph:
             utils.graph_winrates(collected_winrates, args.SGF_FILE)
 
-        # Now fill in variations for everything we need (suggested variations)
+        # Get cursor to first node
         move_num = -1
         cursor = sgf.cursor()
+
+        # Start Leela to analyze variations
         leela.start()
         add_moves_to_leela(cursor, leela)
 
+        # START VARIATIONS ANALYSIS
         while not cursor.atEnd:
             cursor.next()
             move_num += 1
+
+            # Add move to history
             add_moves_to_leela(cursor, leela)
 
+            # Skip if move doesn't need variations analysis
             if move_num not in needs_variations:
                 continue
 
+            # Get stats and move list of current move and next move position
             stats, move_list = needs_variations[move_num]
-            next_game_move = None
+            next_game_move = next_move_pos(cursor)
 
-            if not cursor.atEnd:
-                cursor.next()
-
-                if 'W' in cursor.node.keys():
-                    next_game_move = cursor.node['W'].data[0]
-
-                if 'B' in cursor.node.keys():
-                    next_game_move = cursor.node['B'].data[0]
-
-                cursor.previous()
-
-            do_variations(cursor, leela, stats, move_list, board_size, next_game_move, base_dir, args)
+            # Analyze variations
+            do_variations(cursor, leela, stats, move_list, board_size, next_game_move, base_dir)
             variations_tasks_done += 1
 
-            # save to file results with analyzing variations
             utils.write_to_file(args.save_to_file, 'w', sgf)
-
             refresh_progress_bar()
+
     except:
         traceback.print_exc()
         print("Failure, reporting partial results...", file=sys.stderr)
@@ -557,15 +638,11 @@ if __name__ == '__main__':
         leela.stop()
 
     progress_bar.finish()
-
-    # Save final results into file
     utils.write_to_file(args.save_to_file, 'w', sgf)
-
     time_stop = datetime.datetime.now()
 
     if args.verbosity > 0:
         print("Leela analysis stopped at %s" % time_stop, file=sys.stderr)
         print("Elapsed time: %s" % (time_stop - time_start), file=sys.stderr)
 
-    # delay in case of sequential running of several analysis
     time.sleep(1)
